test(content): split PageContentValidatorTest by rule groups
Split the oversized unit suite into structure, image, and block-field
classes with a shared translator test case.
