+++ docs/03-development/test-architecture-audit.md
@@ -48,7 +48,7 @@ Infrastructure already in good shape:
 | ~~925~~ split | `ExplorerResultsTablePresenter*Test.php` (Columns / Matrix / BoxPlot / HospitalMetrics) | Done |
 | ~~717~~ split | `PageImageContentAnalyzer*Test.php` (ImageBlock / HtmlSnippet) | Done |
 | ~~611~~ split | `DashboardController*Test.php` (Widgets / Scope / Period) | Done |
-| ~577 | `tests/Content/Unit/Page/PageContentValidatorTest.php` | Rule groups |
+| ~~577~~ split | `PageContentValidator*Test.php` (Structure / Image / BlockFields) | Done |
 | ~535 | `tests/Allocation/Integration/Query/AllocationBucketQueryTest.php` | Query variants |
 | ~499 | `tests/Allocation/Functional/Controller/Allocations/ListAllocationsControllerTest.php` | Filters / pagination / access |
